Make CopyToClipboardButton widget, remove python2-compatibility stuff.
